refactor: rename wrong 'useAnimatedNode'
